Create GN-Mark chain before starting EP watcher
In the current code, we are starting the endpoint watcher
and later creating the Globalnet marking chain. In a KIND
deployment this was fine, but in an OCP deployment the endpoint
events are generated even before the iptable chain is programmed
because of which we are seeing errors. This PR fixes it by changing
the order.

Related to: https://github.com/submariner-io/submariner/issues/1749
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>