Fix connection failure with LB mode + VxLAN Cabledriver (#1851)
When using Loadbalancer mode with VxLAN cable driver, it was seen that
connections are sometimes not setup. The reason for this is the missing
vxlan-tunnel interface which was deleted by route-agent driver assuming
that the Gateway node migrated.

The reason why route-agent considers it as Gateway migration is because
the local-endpoint is deleted and then re-created. Ideally, the local
endpoint should not be deleted when the Gateway pod comes up, but the
current checks in the code tries to match the entire backendConfig and
when it finds that something does not match with the endpoint stored in
the datastore it deletes and re-creates the endpoint. This PR fixes this
issue by ensuring that local-endpoint is simply updated and not deleted.

Fixes: https://github.com/submariner-io/submariner/issues/1822
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>