Ignore context linting on fatal()
contextcheck complains about the sequence fatal ->
SetGatewayStatusError -> syncGatewayStatusSafe ->
cleanupStaleGatewayEntries (the last uses a context, and there is a
context in the becameLeader function) but we don't care.

We can't //nolint this in main.go because of the varying behaviour in
the linter between golangci-lint 1.47.3 and 1.50.1, so ignore
contextcheck entirely for main.go for the time being.

Signed-off-by: Stephen Kitt <skitt@redhat.com>