Fix e2e ext-net socket permission denied error
The e2e, external-net tests, are failing in CI and there
seems to be multiple problems. One of them is related to
the port 80 that is currently used by the test which is
the reason for Permission denied error when the pod is
scheduled in a cluster. This PR changes the port to the
default port we use in other e2e tests, which is 1234.

Error seen in pod logs:
socket.error: [Errno 13 Permission denied

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>