Remove Twitter Together automation
We never managed to get the keys setup for this correctly.

This version of the tooling has been archived. The project has moved to:
https://github.com/twitter-together/action.
If we want to prioritize this, we'd need to start over anyway.

Generally, I don't think we get much value from Twitter and I worry
someone who pays for a blue check could impersonate us and link to a
malicious release.

Fixes: #1650
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>