Synchronize all callbacks in the event controller
The controller was using the syncMutex to protect access to the isGatewayNode
field, although not really necessary since the callbacks emanate from the
same syncer on the same thread.

In addition, some handlers employ their own mutex across multiple callbacks,
although unnecessary in most cases. However a hidden caveat with the handler
API is that while the Endpoint-related events (ie LocalEndpointCreated,
TransitionToGateway etc) are single-threaded wrt one another and don't
need synchronization, the Mode-related events emanate from a separate
syncer and thread. So if data is accessed across Endpoint and Node
events then synchronization is needed. Rather then having each handler
implementation be aware of this behavior, it is simpler and safer to
have the event controller use the syncMutex to synchronization access
to all the callbacks and remove the synchronization in the handlers.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>